# Conflicts:
#	common/src/main/java/com/project/hiptour/common/reviews/entity/Review.java
#	common/src/main/java/com/project/hiptour/common/reviews/repository/ReviewRepository.java
#	common/src/main/java/com/project/hiptour/common/reviews/service/ReviewServiceImpl.java
#	sync/src/main/java/com/project/hiptour/sync/application/SyncPlaceCommandHandler.java

+++ common/src/main/java/com/project/hiptour/common/reviews/repository/PlaceRepository.java
@@ -1,12 +1,10 @@
 package com.project.hiptour.common.reviews.repository;
 
-import com.project.hiptour.common.place.Place;
-import com.project.hiptour.common.reviews.entity.Review;
+import com.project.hiptour.common.entity.place.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlaceRepository extends JpaRepository<Place, Long> {
+public interface PlaceRepository extends JpaRepository<Place, Integer> {
     boolean existsByPlaceNameAndAddress1(String placeName, String address1);
 }
